backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
aiosqlite==0.20.0
chromadb==0.5.23
sentence-transformers==3.3.1
anthropic==0.42.0
pypdf==5.1.0
python-docx==1.1.2
unstructured==0.16.11
langchain-text-splitters==0.3.4
sse-starlette==2.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
